fix(auth): 로그인·회원가입 FSD 위반 및 getMe 타이밍 버그 수정
pages 레이어에서 shared/api 직접 호출 제거.
useLogin·useRegister 훅으로 인증 로직을 features로 이동.
login 후 getMe 호출 전 accessToken을 localStorage에 먼저 저장해
Authorization 헤더가 누락되는 버그 수정.
