+++ src/features/auth/index.ts
@@ -1,1 +1,3 @@
 export { useAuthStore } from './model/store'
+export { useLogin } from './login/useLogin'
+export { useRegister } from './register/useRegister'
